LayoutTests/platform/ios-sim-deprecated/plugins/layout-in-beforeload-listener-affects-plugin-loading-expected.txt

CONSOLE MESSAGE: line 27: PASS
